added method call in BigFiveQuestionService to delete all questions before mapping the new questions and saving to the database

+++ server/src/main/java/com/hackathon/server/personalityAPI/PersonalityTestQuestionDTO.java
@@ -44,7 +44,7 @@ public class PersonalityTestQuestionDTO{
         this.question = question;
     }
 
-    public int getPersonalityTestQuestionId() {
+    public long getPersonalityTestQuestionId() {
         return personalityTestQuestionId;
     }
